Incremental commit
- Split compute environments out into two separate environments, one
backed by a manually scaled ASG, the other backed by an ASG that is
scaled in/out via a CloudWatch alarm

- Rename ScaleOutFactor to ScaleFactor, and publish -1/0/1 values to the
metric, allowing the calculation of a scale-in requirement.

- Add second CloudWatch alarm for scale-in, which checks for the job
queue being smaller than the compute environment's capacity.

- Factor SCALE_CAP in when submitting metrics to CloudWatch - should not
attempt add instances to the cluster once the ASG's MAX capacity has
been reached.

- Update inputs for CloudWatch event targets to use new JSON schema.

- Add ComputeEnvironment as an extra dimension on CloudWatch metrics (we
may want to adjust a specific compute environment out of many that are
assigned to the same job queue).

- Miscellaneous refactoring.